requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
atomicwrites==1.3.0
attrs==19.3.0
backports.functools-lru-cache==1.6.1
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
configparser==4.0.2
crayons==0.3.0
Faker==2.0.4
future==0.18.2
idna==2.8
importlib-metadata==0.23
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
py==1.8.0
pyparsing==2.4.5
pytest==5.3.0
python-dateutil==2.8.1
requests==2.22.0
selene==1.0.0a16
selenium==4.0.0a3
six==1.13.0
text-unidecode==1.3
urllib3==1.25.7
wcwidth==0.1.7
webdriver-manager==2.3.0
zipp==0.6.0
